Added PubChem2D fingerprint to data
The respective code is in the notebook 'enrich_dataset_fingerprint'. The
new data set does not contain the CAS anymore, and all the chemical
property features we had before were also dropped. They should be
represented by the fingerprint now. The
functions cannot deal with the new data set yet. This has to be changed
in the next step.