Impl KNN with fingerprint
The mechanics of the KNN algorithm was slightly changed. The distance
matrix is computed differently: we can now supply groupings of features.
Hamming's distance is calculated on each group of features and then
combined with weights 'alpha', which is now a dictionary. If one group
of features consists of exactly one string feature, the Hamming's
distance is calculated on the individual characters. At the moment,
we cannot deal with numerical features, only with categorical ones.